En recuperacion de Cuenta, el boton Agregar otra pregunta es de color dorado. En recuperacion de cuenta, el combo tiene flecha negra deberia ser blanca. En Recuperacion de cuenta, esta permitiendo hacer mas de 3 intentos. En el boton de nota de voz hay rebote al presionar dos veces muy rapido. La hora en los mensajes entrantes se corta al final. En recuperacion de cuenta si se envian las preguntas y tiene una mayuscula, lo toma como incorrecto. Cuando se toca la notificación a veces no lleva la conversación. Algunos textos de las conversaciones quedan en cursiva. Adaptar la vista previa del fondo en conversación a los nuevos cambios del tema. Agregar dialog que notifique cuando la otra persona cancela la solicitud de videollamada estando en una llamada. El badge cuando llega a numero de 100 mensajes solo se ven dos digitos. Android manda solicitud de amistad a iOS. iOS acepta solicitud, se crean las conversaciones con el mensaje sistematico bien en iOS, pero en Android queda mensaje sistematico vacio, solo la burbuja con el tiempo de autodestruccion. Separar tiempo de autodestruccion en mensaje sistematico de llamada perdida. Revisar cunado se bloquea un contacto, algunas veces no se lista en la vista de contactos bloqueados. Cuando uno bloquea contactos, y recupera cuenta los contactos se pierden... no se muestran en ningun lado. Cuando se cierra sesión el trae el ultimo estado del usuario, pero este no se encuentra entre los creados por defecto en bd. Drawer se muestra en la conversacion. En la solicitud de paso de llamada a videollamada colocar el contacto solicita pasar a videollamada.

+++ app/src/main/java/com/naposystems/napoleonchat/repository/sharedRepository/ContactRepositoryShareRepository.kt
@@ -15,20 +15,24 @@ class ContactRepositoryShareRepository @Inject constructor(
     private val messageLocalDataSource: MessageDataSource
 ) : IContractContactRepositoryShare.Repository {
 
-    override suspend fun getContacts(): Boolean {
+    override suspend fun getContacts(state : String, location : Int): Boolean {
         return try {
-            val response = napoleonApi.getContactsByState(Constants.FriendShipState.ACTIVE.state)
+            val response = napoleonApi.getContactsByState(state)
 
             if (response.isSuccessful) {
 
                 val contactResDTO = response.body()!!
 
-                val contacts = ContactResDTO.toEntityList(contactResDTO.contacts)
+                val contacts = if (state == Constants.FriendShipState.BLOCKED.state)
+                    ContactResDTO.toEntityList(contactResDTO.contacts, true)
+                else
+                    ContactResDTO.toEntityList(contactResDTO.contacts)
 
-                val contactsToDelete = contactLocalDataSource.insertOrUpdateContactList(contacts)
+                val contactsToDelete = contactLocalDataSource.insertOrUpdateContactList(
+                    contacts, location
+                )
 
-                if (contactsToDelete.isNotEmpty()) {
-
+                if (contactsToDelete.isNotEmpty() && location == Constants.LocationGetContact.OTHER.location) {
                     contactsToDelete.forEach { contact ->
                         messageLocalDataSource.deleteMessageByType(
                             contact.id,
@@ -37,7 +41,6 @@ class ContactRepositoryShareRepository @Inject constructor(
                         contactLocalDataSource.deleteContact(contact)
                     }
                 }
-
                 true
             } else {
                 Timber.e(response.errorBody()!!.string())
